server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1
psutil==5.9.6
GPUtil==1.4.0
jupyter-client==8.6.0
jupyter-server==2.11.1
nbformat==5.9.2
httpx==0.25.2
authlib==1.3.0
itsdangerous==2.1.2
